Create new app to share the repeated code and logic
