+++ mmdt/mmdt/settings.py
@@ -34,6 +34,7 @@ ALLOWED_HOSTS = ['mmdt.istarvz.com', 'mmdt-dev.thingaha.org', '127.0.0.1']
 INSTALLED_APPS = [
     'polls.apps.PollsConfig',
     'survey.apps.SurveyConfig',
+    'accounts.apps.AccountsConfig',
     'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
