+++ mmdt/templates/accounts/register.html
@@ -16,5 +16,5 @@
       <button type="submit" class="form-button">Register</button>
     </form>
   </div>
-  <p>Already have an account? <a href="{% url 'polls:login' %}">Login here</a>.</p>
+  <p>Already have an account? <a href="{% url 'accounts:login' %}">Login here</a>.</p>
 {% endblock %}
